requirements.txt
streamlit==1.14.0
mysql-connector-python==8.0.31
google-generativeai==0.1.0
pandas==2.0.2
python-dotenv==0.21.1